Board Approval Change Docs & SAP Integration PurposeOfLoan, LoanType, ProjectTypeCoreSector

+++ src/main/java/pfs/lms/enquiry/action/projectproposal/projectdetail/ProjectDetailResource.java
@@ -34,7 +34,7 @@ public class ProjectDetailResource {
     private String policyExposure;
     private String endUseOfFunds;
     private Double fees;
-    private Integer roi;
+    private Double roi;
     private Integer tenorYear;
     private Integer tenorMonths;
     private Integer moratoriumPeriod;
